+++ packages/worker-core/src/index.ts
@@ -76,7 +76,6 @@ export {
 
 export {
   hasTapHeaders,
-  extractIssuerFromKeyid,
   headersToPlainObject,
   verifyTap,
   handleVerification,